+++ core/api_container/server/lambda_store/lambda_launcher/lambda_launcher.go
@@ -13,7 +13,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_consts"
 	"github.com/kurtosis-tech/kurtosis-core/api_container/server/lambda_store/lambda_store_types"
 	"github.com/kurtosis-tech/kurtosis-core/commons"
-	"github.com/kurtosis-tech/kurtosis-core/commons/container_guid_suffix_provider"
+	"github.com/kurtosis-tech/kurtosis-core/commons/current_time_str_provider"
 	"github.com/kurtosis-tech/kurtosis-core/commons/object_labels_providers"
 	"github.com/kurtosis-tech/kurtosis-core/commons/object_name_providers"
 	"github.com/kurtosis-tech/kurtosis-lambda-api-lib/golang/kurtosis_lambda_docker_api"
@@ -167,6 +167,6 @@ func waitUntilLambdaContainerIsAvailable(ctx context.Context, client kurtosis_la
 }
 
 func newLambdaGUID(lambdaID lambda_store_types.LambdaID) lambda_store_types.LambdaGUID {
-	suffix := container_guid_suffix_provider.GetContainerGUIDSuffix()
+	suffix := current_time_str_provider.GetCurrentTimeStr()
 	return lambda_store_types.LambdaGUID(string(lambdaID) + "_" + suffix)
 }