+++ core/docs/changelog.md
@@ -2,7 +2,7 @@
 ### Features
 * Added a new `interactive-repl` value to the `container-type` label
 * Added `ForInteractiveREPLContainer` functions to the `EnclaveObjectLabelsProvider` and `EnclaveObjectNameProvider`
-* Added a `GetContainerGUIDSuffix` function to centralize generation of container GUID suffixes
+* Added a `GetCurrentTimeStr` function to centralize generation of container GUID suffix strings
 
 # 1.23.3
 ### Features